requirements.txt
numpy>=1.24
pandas>=1.3
opencv-contrib-python==4.10.0.84
matplotlib
colorlog
Pillow
tqdm
